Epic Loot 0.7.10 - RC1
  * Increased epic and legendary drop rates at low tiers
  * Wishbone is now Epic rarity
  * Dragon's Tears and Yagluth Things can now be converted to legendary runestones or sacrificed for legendary crafting mats
  * Moder and Yagluth trophies sacrifice for more runestones (2, 4 instead of 1, 3 respectively)
  * Bosses drop more items overall, especially Moder and Yagluth
  * Weights in config files are now floats instead of ints
  * Magic Effect: Luck, increase chance for higher rarity items
  * Magic Effect names shortened, detailed descriptions added to compendium

+++ EpicLoot/Adventure/AdventureDataConfig.cs
@@ -44,8 +44,8 @@ namespace EpicLoot.Adventure
         public int ForestTokenGamblesCount;
         public int IronBountyGamblesCount;
         public int GoldBountyGamblesCount;
-        public int[] GambleRarityChance = new int[5];
-        public int[][] GambleRarityChanceByRarity = { new int[5], new int[5], new int[5], new int[5] };
+        public float[] GambleRarityChance = new float[5];
+        public float[][] GambleRarityChanceByRarity = { new float[5], new float[5], new float[5], new float[5] };
         public float ForestTokenGambleCoinsCost = 1.0f;
         public int ForestTokenGambleCostMin = 5;
         public int ForestTokenGambleCostMax = 10;
